+++ internal/app/models.go
@@ -84,7 +84,7 @@ type SummaryStore interface {
 	DeleteSummary(recipeID string) error
 	AddMashTemp(id string, temp float32, notes string) error
 	AddRast(id string, temp float32, duration float32, notes string) error
-	AddLauternNotes(id, notes string) error
+	AddLauternNotes(id, notes string, duration float32) error
 	AddHopping(id string, name string, amount float32, alpha float32, duration float32, notes string) error
 	AddVolumeBeforeBoil(id string, amount float32, notes string) error
 	AddVolumeAfterBoil(id string, amount float32, notes string) error
@@ -96,7 +96,7 @@ type SummaryStore interface {
 	AddDryHopStart(id string, name string, amount, alpha float32, notes string) error
 	AddDryHopEnd(id string, name string, durationHours float32) error
 	AddPreBottlingVolume(id string, volume float32) error
-	AddBottling(id string, carbonation, alcohol, sugar, water, temp, vol float32, sugarType, notes string) error
+	AddBottling(id string, carbonation, alcohol, sugar, water, temp, vol, time_min float32, sugarType, notes string) error
 	AddSummarySecondary(id string, days int, notes string) error
 	AddFinishedTime(id string, t time.Time) error
 	AddEvaporation(id string, amount float32) error