+++ internal/db_migrations/migrations/000013_add_lautern_duration.up.sql
@@ -1,0 +1,2 @@
+ALTER TABLE "summaries"
+ADD COLUMN lautern_duration REAL;