+++ internal/routers/lautern/lautern.go
@@ -10,10 +10,11 @@ import (
 )
 
 type LauternRouter struct {
-	TLStore      TimelineStore
-	SummaryStore SummaryStore
-	Store        RecipeStore
-	Timer        Timer
+	TLStore         TimelineStore
+	SummaryStore    SummaryStore
+	Store           RecipeStore
+	Timer           Timer
+	LauternRestTime int
 }
 
 // RegisterRoutes adds routes to the web server
@@ -35,9 +36,9 @@ func (r *LauternRouter) addTimelineEvent(id, message string) error {
 }
 
 // addSummaryLauternNotes adds lautern notes to the summary
-func (r *LauternRouter) addSummaryLauternNotes(id, notes string) error {
+func (r *LauternRouter) addSummaryLauternNotes(id, notes string, duration float32) error {
 	if r.SummaryStore != nil {
-		return r.SummaryStore.AddLauternNotes(id, notes)
+		return r.SummaryStore.AddLauternNotes(id, notes, duration)
 	}
 	return nil
 }
@@ -70,7 +71,7 @@ func (r *LauternRouter) getLauternHandler(c echo.Context) error {
 		"RecipeID":         id,
 		"MashOutTemp":      re.Mashing.MashOutTemperature,
 		"Hops":             re.Hopping.Hops,
-		"RestTime":         15,
+		"RestTime":         r.LauternRestTime,
 		"StartClickedOnce": started,
 		"Stopped":          stopped,
 	})
@@ -91,7 +92,7 @@ func (r *LauternRouter) postLauternHandler(c echo.Context) error {
 	if err != nil {
 		log.Error().Str("id", id).Err(err).Msg("could not add timeline event")
 	}
-	err = r.addSummaryLauternNotes(id, req.Notes) //TODO: handle real duration also
+	err = r.addSummaryLauternNotes(id, req.Notes, req.RealDuration)
 	if err != nil {
 		log.Error().Err(err).Str("id", id).Msg("Failed to add lautern notes to summary")
 	}
@@ -103,7 +104,7 @@ func (r *LauternRouter) getLauternTimestamp(c echo.Context) error {
 	if id == "" {
 		return common.ErrNoRecipeIDProvided
 	}
-	dur := 15 * time.Minute //TODO: Make this configurable
+	dur := time.Duration(r.LauternRestTime) * time.Minute
 	return r.Timer.HandleStartTimer(c, id, dur, "lautern")
 }