+++ internal/summary/memory/memory.go
@@ -86,14 +86,18 @@ func (s *SummaryMemoryStore) AddRast(id string, temp float32, duration float32, 
 }
 
 // AddLauternNotes adds lautern notes to the summary
-func (s *SummaryMemoryStore) AddLauternNotes(id, notes string) error {
+func (s *SummaryMemoryStore) AddLauternNotes(id, notes string, duration float32) error {
 	s.lock.Lock()
 	defer s.lock.Unlock()
 	sum, err := s.getSummary(id)
 	if err != nil {
 		return err
 	}
-	sum.LauternInfo = notes
+	if sum.LauternInfo == nil {
+		sum.LauternInfo = &summary.LauternInfo{}
+	}
+	sum.LauternInfo.Notes = notes
+	sum.LauternInfo.Duration = duration
 	return nil
 }
 
@@ -312,7 +316,7 @@ func (s *SummaryMemoryStore) AddPreBottlingVolume(id string, volume float32) err
 }
 
 // AddBottling adds a summary of the bottling
-func (s *SummaryMemoryStore) AddBottling(id string, carbonation, alcohol, sugar, water, temp, vol float32, sugarType, notes string) error {
+func (s *SummaryMemoryStore) AddBottling(id string, carbonation, alcohol, sugar, water, temp, vol, time_min float32, sugarType, notes string) error {
 	s.lock.Lock()
 	defer s.lock.Unlock()
 	sum, err := s.getSummary(id)
@@ -330,6 +334,7 @@ func (s *SummaryMemoryStore) AddBottling(id string, carbonation, alcohol, sugar,
 	sum.BottlingInfo.Temperature = temp
 	sum.BottlingInfo.VolumeBottled = vol
 	sum.BottlingInfo.Notes = notes
+	sum.BottlingInfo.Time = time_min
 	return nil
 }