+++ internal/summary/printer/markdown/markdown_test.go
@@ -64,7 +64,10 @@ func TestPrint(t *testing.T) {
 						},
 					},
 				},
-				LauternInfo: "lautern info 1\nlautern info 2",
+				LauternInfo: &summary.LauternInfo{
+					Notes:    "lautern info 1\nlautern info 2",
+					Duration: 90,
+				},
 				HoppingInfo: &summary.HoppingInfo{
 					VolBeforeBoil: &summary.VolMeasurement{Volume: 13.10, Notes: "notes4"},
 					HopInfos: []*summary.HopInfo{
@@ -109,6 +112,7 @@ func TestPrint(t *testing.T) {
 					Water:           0.5,
 					Alcohol:         5.86,
 					VolumeBottled:   10.00,
+					Time:            90,
 					Notes:           "notes19",
 				},
 				SecondaryFermentationInfo: &summary.SecondaryFermentationInfo{Days: 5, Notes: "notes20"},
@@ -131,6 +135,8 @@ The following summary was generated on %s
 
 
 ## Lautern
+
+Duration: 90.00 min
 
 lautern info 1
 lautern info 2
@@ -181,6 +187,8 @@ Date | SG | Final | Notes
 
 
 ## Bottling
+
+Duration: 90.00 min
 
 - **Volume in tank**: 8.80L